+++ src/libraries/System.Private.CoreLib/src/ILLink/ILLink.Descriptors.Shared.xml
@@ -43,6 +43,7 @@
     <type fullname="System.Runtime.InteropServices.CustomMarshalers.*" />
     <!-- Workaround for https://github.com/mono/linker/issues/378 -->
     <type fullname="System.Runtime.InteropServices.IDispatch" />
+    <type fullname="System.Runtime.InteropServices.IMarshal" />
     <type fullname="Internal.Runtime.InteropServices.IClassFactory2" />
     <type fullname="System.Threading.ThreadPoolBoundHandle">
       <!-- Workaround to keep .interfaceimpl even though this type
